Added menu for stations
Had to change how the select update function updated its variables, as it now changes the select-pressed variable to 0 after it runs all needed functions, rather than letting it be changed on button released.

Now the menu update function is called on screen/canvas proc, through the menu select function.

This allows me to no longer initialize the "first planet" in the init function of main.  It also allows the menus to update their info when the player opens them, so the menu contains more up to date information.

The station menu just displays a list of names for each station, and resizes the menu to depend on the # of stations.  So far it doesn't do anything yet, no trading or anything.

Planned stations:

Trade: buy and sell

Civilian: Bartering and/or missions

Military: Don't attack this, unless you are strong, and like loot

Warp: Spend some resource to go to the next sector, that isn't antimatter.